requirements.txt
# ThinkingSDK E2E Testing Requirements

# Core ThinkingSDK Client
thinking_sdk_client
keyring
#-e ../thinkingSDK/thinking_sdk_client

# Testing and GitHub Integration
aiohttp>=3.8.0
requests>=2.28.0
python-dotenv>=1.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# GitHub API and Git Operations
PyGithub>=1.59.0
GitPython>=3.1.32

# Async and Concurrency
asyncio-throttle>=1.0.2

# Environment and Process Management
psutil>=5.9.0

# Optional: Rich console output for better test visualization
rich>=13.0.0
